requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
celery-pool-asyncio==0.2.0
certifi==2024.12.14
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.0
daphne==4.1.2
debugpy==1.8.5
Django==5.1.5
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-crontab==0.7.1
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Flask==3.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.4.2
MarkupSafe==3.0.2
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-crontab==3.2.0
watchgod==0.8.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytx==0.5.10
pytz==2024.2
rarfile==4.2
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
tabulate==0.9.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.3
zope.interface==7.2
